refactor: address PR review feedback on diff viewer
- Split session_diffs_screen.dart (321 lines, over the documented
  250-line app file limit) by component: SessionDiffsBody moves to its
  own file, the expanded-file content sliver and the error view become
  focused widgets in widgets/. The screen file keeps only the
  BlocProvider shell, Scaffold/AppBar, and the stats title. All feature
  files are now <= 250 lines. The collapse/scroll compensation logic
  intentionally stays as private methods on the body State (it owns
  that state), per the repo ownership-boundary rule - no helper/mixin
  extraction.

- Harden _isHeaderPinnedAtTop per review: guard RenderBox.attached
  before coordinate conversion, and compute the header offset directly
  in the scrollable's coordinate space via
  localToGlobal(ancestor: scrollableBox) instead of subtracting two
  global offsets.
